Bullshit

+++ common/pid.h
@@ -24,19 +24,21 @@
 
 #pragma once
 
+#include <stdint.h>
+
 #pragma pack(push,4)
 
 struct process_id {
   unsigned ip;
   short port;
-  unsigned short pid;
+  uint32_t pid;
   int utime;
 };
 
 struct process_id_ext {
   unsigned ip;
   short port;
-  unsigned short pid;
+  uint32_t pid;
   int utime;
   int actor_id;
 };
